Add list albums method and add to app.rb
